+++ circuit_maintenance_parser/__init__.py
@@ -9,6 +9,7 @@ from .parsers.lumen import ParserLumen
 from .parsers.megaport import ParserMegaport
 from .parsers.ntt import ParserNTT
 from .parsers.packetfabric import ParserPacketFabric
+from .parsers.telstra import ParserTelstra
 from .parsers.zayo import ParserZayo
 
 
@@ -18,6 +19,7 @@ SUPPORTED_PROVIDER_PARSERS = (
     ParserMegaport,
     ParserNTT,
     ParserPacketFabric,
+    ParserTelstra,
     ParserZayo,
     ICal,
 )