+++ circuit_maintenance_parser/output.py
@@ -5,7 +5,8 @@ See: https://tools.ietf.org/html/draft-gunter-calext-maintenance-notifications-0
 
 import json
 from enum import Enum
-from typing import List, Union
+
+from typing import List, Optional
 
 from pydantic import BaseModel, validator, StrictStr, StrictInt
 
@@ -135,7 +136,8 @@ class Maintenance(BaseModel):
     organizer: StrictStr
 
     # Non mandatory attributes
-    stamp: Union[StrictInt, None] = None
+
+    stamp: Optional[StrictInt] = None
     uid: StrictStr = "0"
     sequence: StrictInt = 1
     summary: StrictStr = ""