+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - #16 - changed `MaintenanceNotification.raw` from `str` to `bytes`
 - #17 - added **Lumen** parser and refactor HTML parser `process` method
+- #18 - added **Telstra** parser and made `Maintenance.stamp` attribute optional
 
 ## v1.0.2 - 2021-05-05